% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_JumpBack}
\alias{assert_JumpBack}
\title{Echo personalized error message for jump back}
\usage{
assert_JumpBack(x)
}
\arguments{
\item{x}{User input (string)}
}
\description{
Echo personalized error message for jump back
}
\details{
Checks if input is a string and either yes or no
}
\keyword{internal}
