% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_valid_parameters}
\alias{check_valid_parameters}
\title{Check valid parameters of function simHostSwitch}
\usage{
check_valid_parameters(data)
}
\arguments{
\item{data}{A list of arguments}
}
\description{
Check valid parameters of function simHostSwitch
}
\details{
Uses checkmate to evaluate arguments (either default or provided by the user) from simHostSwitch function
}
\keyword{internal}
