% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{createPlotInput_sim_pInd}
\alias{createPlotInput_sim_pInd}
\title{Create input data of simulated individuals for plots}
\usage{
createPlotInput_sim_pInd(x)
}
\arguments{
\item{x}{A nested list}
}
\description{
Create input data of simulated individuals for plots
}
\details{
Flattens a nested list; https://stackoverflow.com/questions/8139677/how-to-flatten-a-list-to-a-list-without-coercion
}
\keyword{internal}
