% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{createPlotInput_sim_which}
\alias{createPlotInput_sim_which}
\title{Create input data of simulated jumped/survived individuals for plots}
\usage{
createPlotInput_sim_which(x)
}
\arguments{
\item{x}{A list of simulated jumped/survived parasite phenotypes for each generation}
}
\description{
Create input data of simulated jumped/survived individuals for plots
}
\details{
Converts list to dataframe and stack it
}
\keyword{internal}
