% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{flatten2}
\alias{flatten2}
\title{Flatten a nested list}
\usage{
flatten2(x)
}
\arguments{
\item{x}{A nested list}
}
\description{
Flatten a nested list
}
\details{
Flattens a nested list; https://stackoverflow.com/questions/8139677/how-to-flatten-a-list-to-a-list-without-coercion
}
\keyword{internal}
