% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parli.R
\docType{data}
\name{parli}
\alias{parli}
\title{Consumer life cycle parameters to simulate host switches}
\format{
An object of class "list".
}
\usage{
data(parli)
}
\description{
Data derived from real world experiments (see references). Data includes
life cycle parameters for a wildlife ecology, agricultural
and biomedical arena  to simulate host switches.
}
\examples{
data(parli)
knitr::kable(parli$Cephaloleia) # tibble::as.tibble(parli$Cephaloleia)

}
\references{
Trivellone V, Araujo SBL and Panassiti B
(2021) HostSwitch: An R Package to Simulate
the Extent of Host-Switching by a Consumer
}
\keyword{datasets}
