% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostSwitch.R
\name{simHostSwitch}
\alias{simHostSwitch}
\title{Simulate host switches by consumers}
\usage{
simHostSwitch(
  K = 100,
  b = 10,
  mig = 0.01,
  sd = 0.2,
  sigma = 1,
  pRes_min = 1,
  pRes_max = 10,
  n_generation = 200,
  jump_back = "no",
  seed = NULL,
  n_sim = 1
)
}
\arguments{
\item{K}{Carrying capacity, numeric value (min=0, max=1000)}

\item{b}{Average number of offspring each consumer can have (birth rate), numeric value (min=0, max=K)}

\item{mig}{Cut off for migration, individuals below cutoff jump, numeric value (min=0, max=1)}

\item{sd}{Standard deviation for mutation, numeric value (min=0, max=10)}

\item{sigma}{Standard deviation for selection, numeric value (min=0, max=10)}

\item{pRes_min}{Initial value, smallest phenotype of resource (original host) and consumer, numeric value (min=0, max=pRes_max)}

\item{pRes_max}{Initial value, maximum phenotype of resource (original host) and consumer, numeric value (min=pRes_min, max=100)}

\item{n_generation}{Number of generations, numeric value (min=1, max=50000)}

\item{jump_back}{Options for consumers that do not survive on the new host. If "yes" the consumer(s) jump back to the current host and will be considered in the selective pressure and reproduction stage for the n+1 generation, if "no" (default) it dies on the new host.}

\item{seed}{Random number to ensure reproducible plots, positive integer}

\item{n_sim}{Number of simulations, positive integer (min=1, max = 50000)}
}
\value{
An object of class HostSwitch
}
\description{
Simulate host switches by consumers
}
\details{
This function simulates the number of host switches by the population of a consumer. Results are stored to a HostSwitch object, to make use of summary and plotting functions in the HostSwitch package. The HostSwitch object includes the following simulated quantities are: $pRes_sim (all the optimal phenotypes favored by the selected new hosts), $pRes_new_sim (new resource), $pInd and of individual consumer. These simulated quantities of interest are available for each generation step and can be used for summary statistics or plots.
}
\examples{
m1 = simHostSwitch() # default values for arguments
m1
}
