% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostSwitch.R
\name{summaryHostSwitch}
\alias{summaryHostSwitch}
\title{Summary statistics of HostSwitch simulation}
\usage{
summaryHostSwitch(HostSwitch_simulated_quantities)
}
\arguments{
\item{HostSwitch_simulated_quantities}{An object created by \code{\link{simHostSwitch}}}
}
\value{
Summary of HostSwitch simulations
}
\description{
Summary statistics of HostSwitch simulation
}
\details{
This function generates summary statistcs for HostSwitch simulations.
}
\examples{
m1 = simHostSwitch(n_sim=100) # except n_sim, default values for arguments
summaryHostSwitch(m1)
}
