% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHostSwitch.R
\docType{class}
\name{testHostSwitch-class}
\alias{testHostSwitch-class}
\title{Set class and method}
\description{
This is a build-time dependency on methods, as opposed to a run-time
dependency, thus requiring the importFrom tag to avoid a NOTE when checking
the package on CRAN.
}
\keyword{internal}
