% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{createPlotInput_sim_pInd}
\alias{createPlotInput_sim_pInd}
\title{Create input data of simulated individuals for plots}
\usage{
createPlotInput_sim_pInd(x)
}
\arguments{
\item{x}{A list of simulated parasite phenotypes for each generation}
}
\description{
Create input data of simulated individuals for plots
}
\details{
Converts list to dataframe and stack it
}
\keyword{internal}
