% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHostSwitch}
\alias{plotHostSwitch}
\title{HostSwitch standard plot}
\usage{
plotHostSwitch(x)
}
\arguments{
\item{x}{an object of created by \code{\link{simHostSwitch}}}
}
\description{
This function creates a standard plot for the simulated host switches
}
\details{
The function creates a plot with host switches by a parasite. X axis shows the number of generations, y axis the considered phenotpye range.

Red squares show the phenotype of the resource (original host), green squares the phenotype of the new host and black dots individual parasites.
}
