% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHostSwitch}
\alias{plotHostSwitch}
\title{HostSwitch standard plot}
\usage{
plotHostSwitch(HostSwitch_simulated_quantities, n_sim = 1)
}
\arguments{
\item{HostSwitch_simulated_quantities}{The object HostSwitch created by \code{\link{simHostSwitch}}}

\item{n_sim}{simulation number plotted. The "n_sim" values possible are saved in the HostSwitch object, the plot from the first simulation is shown as default (n_sim = 1).An error message "Error in FUN(X[[i]], ...) : subscript out of bounds" will be returned, if the called simulation is not present in the object HostSwitch.}
}
\description{
This function creates a standard plot for the simulated host switches
}
\details{
The function plots the simulation for the dispersal, colonization (successful host switch), and reproduction of a consumer. The X-axis shows the number of generations, the y-axis the phenotype values of the consumer.

Black dots are the phenotype values after each event of reproduction. The green squares represent the value of phenotype favored by a new host proposed at each generation. The red squares are the phenotype values favored by the current host (=the initial one and the host after successful colonization).
}
\examples{
m1 = simHostSwitch(n_sim=100) # except n_sim, default values for arguments

## The results of the first simuation´(n_sim=1) are plotted as default
plotHostSwitch(m1)

## The 50th simulation of the model is plotted
plotHostSwitch(m1,n_sim=50)

}
