% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyPlots.R
\name{shinyHostSwitch}
\alias{shinyHostSwitch}
\title{Survival probability of parasite}
\usage{
shinyHostSwitch()
}
\description{
Survival probability of parasite
}
\details{
The function creates a shiny plot with host switches by a parasite.
}
