% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyPlots.R
\name{shinyHostSwitch}
\alias{shinyHostSwitch}
\title{Shiny plot of survival probability of parasite after host switch}
\usage{
shinyHostSwitch()
}
\description{
Shiny plot of survival probability of parasite after host switch
}
\details{
The function creates a shiny plot with host switches by a parasite.
}
