% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyPlots.R
\name{shinyHostSwitch}
\alias{shinyHostSwitch}
\title{Shiny-plot of consumer host-switching}
\usage{
shinyHostSwitch()
}
\description{
Shiny-plot of consumer host-switching
}
\details{
The function creates an interactive web-based front-end using Shiny App. The Shiny-Plot simulates dispersal and colonization events of a consumer on novel hosts.

The parameter values included in the function \code{\link{simHostSwitch}} can be modified by the slider bars. The refresh button on the top left (Refresh simulation) plots the new simulation.

Black dots are the phenotype values of the Consumer after each event of reproduction. The green squares represent the value of phenotype values of the Consumer favored by the novel host proposed at each generation. The red squares are the phenotype values favored by the current host.

The total number of tentative dispersal by the parasites and successful host switches are calculated below the plot.
}
\examples{

## Run shiny app locally
 shinyHostSwitch()

}
