% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostSwitch.R
\name{simHostSwitch}
\alias{simHostSwitch}
\title{Simulate host switches by consumers}
\usage{
simHostSwitch(
  K = 100,
  b = 10,
  mig = 0.01,
  sd = 0.2,
  sigma = 1,
  pRes_min = 1,
  pRes_max = 10,
  n_generations = 200,
  jump_back = "no",
  seed = NULL,
  n_sim = 1
)
}
\arguments{
\item{K}{Carrying capacity, positive integer (min=1, max=1000), default value: 100}

\item{b}{Average number of offspring each consumer can have (birth rate), numeric value (min=0, max=K), default value: 10}

\item{mig}{Cut off for migration, individuals below cutoff jump, numeric value (min=0, max=1), default value: 0.01}

\item{sd}{Standard deviation for mutation, numeric value (min=0, max=10), default value: 0.2}

\item{sigma}{Standard deviation for selection, numeric value (min=0, max=10), default value: 1}

\item{pRes_min}{Initial value, smallest phenotype of resource (original host) and consumer, numeric value (min=0, max=pRes_max), default value: 1}

\item{pRes_max}{Initial value, maximum phenotype of resource (original host) and consumer, numeric value (min=pRes_min, max=100), default value: 10}

\item{n_generations}{Number of generations, positive integer (min=1, max=50000), default value: 200}

\item{jump_back}{Options for consumers that do not survive on the new host. If "yes" the consumer(s) jump back to the current host and will be considered in the selective pressure and reproduction stage for the n+1 generation, if "no" (default) it dies on the new host.}

\item{seed}{Random number to ensure reproducible plots, positive integer (>0), default value: NULL}

\item{n_sim}{Number of simulations, positive integer (min=1, max = 50000), default value: 1}
}
\value{
An object of class HostSwitch
}
\description{
Simulate host switches by consumers
}
\details{
This function simulates the number of host switches by the population of a consumer. Results are stored to an object of class \sQuote{HostSwitch},
to make use of summary and plotting functions in the HostSwitch package.\cr\cr
The object of class \sQuote{HostSwitch} includes the following simulated quantities:
\describe{
  \item{}{\bold{$pRes_sim}: optimal phenotypes favored by the \emph{current hosts}}
  \item{}{\bold{$pRes_new_sim}: phenotypes of \emph{new resources}}
  \item{}{\bold{$pInd}: phenotypes of \emph{individual consumer}}
}
These simulated quantities of interest are available for each generation step and can be used for summary statistics \code{\link{summaryHostSwitch}} or plots \code{\link{plotHostSwitch}}.\cr

Note: One important aspect of simHostswitch is the calculation of the survival probability of the consumer on the current and new resources.
Here the argument sigma defines the niche breadth and influences survival probability. For more information see \code{\link{survivalProbability}}.
}
\examples{
m1 = simHostSwitch() # using default values for arguments

\dontrun{
simHostSwitch(sigma=100)}

}
\seealso{
\code{\link{survivalProbability}}, \code{\link{summaryHostSwitch}}, \code{\link{plotHostSwitch}}
}
