% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostSwitch.R
\name{simHostSwitch}
\alias{simHostSwitch}
\title{Simulate host switches by parasites}
\usage{
simHostSwitch(
  K = 100,
  b = 10,
  mig = 0.01,
  sd = 0.2,
  sigma = 1,
  pRes_min = 1,
  pRes_max = 10,
  n_generation = 200,
  jump_back = "no",
  seed = NULL
)
}
\arguments{
\item{K}{Carrying capacity}

\item{b}{Birth rate}

\item{mig}{Migration probability}

\item{sd}{Standard deviation for mutation}

\item{sigma}{Standard deviation for selection}

\item{pRes_min}{Initial value, smallest phenotype of resource (original host) and parasite}

\item{pRes_max}{Initial value, maximum phenotype of resource (original host) and parasite}

\item{n_generation}{Number of generations}

\item{jump_back}{Allow parasites to jump back to previous resource}

\item{seed}{Random number to ensure reproducible plots}
}
\value{
A list with simulated quantities of interest: which can be used for summary statistics or plots. Quantities of interests are phenotpyes of resource (original host, 'pRes_sim'), new resource (new host, 'pRes_new_sim') and of individual parasites ('pInd'). These simulated quantities of interest are available for each generation step.
}
\description{
Simulate host switches by parasites
}
\details{
The use of this function is to simulate host switches by parasites
}
\examples{
simHostSwitch(K=100,b=10, mig=0.01, sd=0.2, sigma=1, pRes_min=1, pRes_max=10, n_generation=200)
}
