% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostSwitch.R
\name{summaryHostSwitch}
\alias{summaryHostSwitch}
\title{Summary statistics of HostSwitch simulation}
\usage{
summaryHostSwitch(HostSwitch_simulated_quantities)
}
\arguments{
\item{HostSwitch_simulated_quantities}{An object created by \code{\link{simHostSwitch}}}
}
\value{
An object of class HostSwitch
}
\description{
Summary statistics of HostSwitch simulation
}
\details{
This function generates summary statistcs for HostSwitch simulations.
}
\examples{
HostSwitch_simulated_quantities = simHostSwitch(K=100,b=10, mig=0.01, sd=0.2, sigma=1, pRes_min=1, pRes_max=10, n_generation=200,iter=100)
summaryHostSwitch(HostSwitch_simulated_quantities)
}
