% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostSwitch.R
\name{summaryHostSwitch}
\alias{summaryHostSwitch}
\title{Summary statistics of HostSwitch simulation}
\usage{
summaryHostSwitch(HostSwitch_simulated_quantities, warmup = 1)
}
\arguments{
\item{HostSwitch_simulated_quantities}{An object created by \code{\link{simHostSwitch}}}

\item{warmup}{Number of warmup steps to be excluded from summary statstistics, see details. Possible value are NULL or positive integer (min=1,max=50). Default value = 1}
}
\value{
Summary of HostSwitch simulations
}
\description{
Summary statistics of HostSwitch simulation
}
\details{
This function generates summary statistcs for HostSwitch simulations.
Warmup represents the initial condition and is defined as an adaptation stage of the simulation model. The initial condition corresponds to the number of
generations (n_generations): warmup = 1 means that the generation at time 0 is excluded from summary; warmup = 2 means generations at times 0 and 1 are excluded and so on.
If warmup = NULL all generations are considered for summary statistics, i.e. initial condition is not considered.
}
\examples{
m1 = simHostSwitch(n_sim=100) # except n_sim, default values for arguments
summaryHostSwitch(m1)
}
