% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HostSwitch.R
\name{survivalProbability}
\alias{survivalProbability}
\title{Survival probability of the consumer in a novel resource (host)}
\usage{
survivalProbability(pInd, pOpt, sigma)
}
\arguments{
\item{pInd}{Phenotype of ith consumer attempting to disperse in a novel resource}

\item{pOpt}{The optimum phenotype the consumer should have to maximize the colonization success}

\item{sigma}{Standard deviation of the niche breadth}
}
\value{
The survival probability of the consumer
}
\description{
Survival probability of the consumer in a novel resource (host)
}
\details{
This function calculates the survival probability of individual consumers that attempt dispersal in a new host. It is the core function of \code{\link{simHostSwitch}}.
The function is based on the probability density function of a normal distribution. By ignoring the normalization constant \eqn{(1/sqrt(2*pi)*sigma)}, it provides then the survival probability.
}
\examples{
## Example 1a - The ith consumer has the phenotype that maximize its
## colonization success on the new host, then pInd is equal to pOpt (pInd = pOpt),
## and the survival probability is 1.
survivalProbability(pInd=5,pOpt=5,sigma=1)

## Example 1b - Increasing |pInd-pOpt| the survival probability decreases
survivalProbability(pInd=5,pOpt=30,sigma=1)

## Example 1c - Give a |pInd-pOpt|> 1, increases sigma the survival probability increases
survivalProbability(pInd=5,pOpt=30,sigma=1)


}
