% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHostSwitch.R
\name{testHostSwitch}
\alias{testHostSwitch}
\title{Test for differences of HostSwitch simulations}
\usage{
testHostSwitch(
  simulated_quantities1,
  simulated_quantities2,
  parameter,
  test,
  plot = FALSE
)
}
\arguments{
\item{simulated_quantities1}{An object created by \code{\link{simHostSwitch}}}

\item{simulated_quantities2}{An object created by \code{\link{simHostSwitch}}}

\item{parameter}{Quantity of interest, possible values are:
\itemize{
  \item "j" for \strong{j}umps of parasites from resource to new resource
  \item "s" for \strong{s}uccessful host switches
  \item "d" for \strong{d}istance between parasite and new host phenotypes if host switch successful
}}

\item{test}{Statistical test, available tests are:
\itemize{
  \item ''t'' for \strong{t}-test
}}

\item{plot}{If \emph{TRUE}, a boxplot is drawn}
}
\value{
An object of class HostSwitch
}
\description{
Test for differences of HostSwitch simulations
}
\details{
This function can be used to compare HostSwitch simulations with different settings and test for differences.
}
\examples{
m1 = simHostSwitch(n_generation=100,iter=100)
m2 = simHostSwitch(n_generation=50,iter=50)
testHostSwitch(simulated_quantities1=m1,simulated_quantities2=m2,parameter="j",test="t",plot=TRUE)
}
