% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHostSwitch.R
\name{testHostSwitch}
\alias{testHostSwitch}
\title{Test for differences of HostSwitch simulations}
\usage{
testHostSwitch(x1, x2, parameter, test, plot = FALSE)
}
\arguments{
\item{x1}{An object created by \code{\link{simHostSwitch}}}

\item{x2}{An object created by \code{\link{simHostSwitch}}}

\item{parameter}{Quantity of interest, possible values are:
\itemize{
  \item "s" for \strong{s}uccessful host switches
  \item "d" for \strong{d}istance between parasite and new host phenotypes if host switch successful
}}

\item{test}{Statistical test, available tests are:
\itemize{
  \item ''t'' for \strong{t}-test
}}

\item{plot}{If \emph{TRUE}, a boxplot is drawn}
}
\value{
An object of class HostSwitch
}
\description{
Test for differences of HostSwitch simulations
}
\details{
This function can be used to compare HostSwitch simulations with different settings and test for differences.
}
\examples{
HostSwitch_simulated_quantities = simHostSwitch(K=100,b=10, mig=0.01, sd=0.2, sigma=1, pRes_min=1, pRes_max=10, n_generation=200,iter=100)
testHostSwitch(HostSwitch_simulated_quantities)
}
